requirements.txt
streamlit
viennarna
biopython
rdkit
pandas
scikit-learn
